Updates
I added checks to the functions to make sure that a user cannot skip through sections when prompted to enter their name, or player names. I also added a check that make sures that the user enters the first and last name of the nba player. Last thing done is I created a function that compares the total points each player has and outputs which player had more points at the end of that season.